+++ modules/core/test/adapter/resources/texture.spec.ts
@@ -203,111 +203,52 @@ test('Texture#format creation with data', async t => {
   t.end();
 });
 
-test.skip('Texture#setParameters', t => {
-  const texture = webglDevice.createTexture({});
-  t.ok(texture instanceof Texture, 'Texture construction successful');
+test('Texture#copyExternaImage', async t => {
+  for (const device of await getTestDevices()) {
+    const texture = device.createTexture({
+      width: 2,
+      height: 1,
+      format: 'rgba8unorm',
+      mipmaps: false
+    });
 
-  testSamplerParameters({t, texture, parameters: SAMPLER_PARAMETERS});
+    if (device.info.type === 'webgl') {
+      t.deepEquals(
+        webglDevice.readPixelsToArrayWebGL(texture),
+        new Uint8Array(8),
+        `${device.info.type} Pixels are initially empty`
+      );
+    }
 
-  /*
-  // Bad tests
-  const parameter = GL.TEXTURE_MAG_FILTER;
-  const value = GL.LINEAR_MIPMAP_LINEAR;
-  texture.setParameters({
-    [parameter]: value
-  });
-  const newValue = getParameter(texture, GL.TEXTURE_MAG_FILTER);
-  t.equals(newValue, value,
-    `Texture.setParameters({[${device.getGLKey(parameter)}]: ${device.getGLKey(value)}})`);
-  */
+    // data: canvas
+    if (typeof document !== 'undefined') {
+      const canvas = document.createElement('canvas');
+      canvas.width = 2;
+      canvas.height = 1;
+      const ctx = canvas.getContext('2d');
+      ctx.fillRect(0, 0, 2, 1);
+      texture.copyExternalImage({image: canvas});
 
-  texture.destroy();
-  t.ok(texture instanceof Texture, 'Texture delete successful');
-
-  t.end();
-});
-
-test.skip('WebGL2#Texture setParameters', t => {
-  const texture = webglDevice.createTexture({});
-  t.ok(texture instanceof Texture, 'Texture construction successful');
-
-  testSamplerParameters({t, texture, parameters: SAMPLER_PARAMETERS_WEBGL2});
-
-  texture.destroy();
-  t.ok(texture instanceof Texture, 'Texture delete successful');
-
-  t.end();
-});
-
-test.skip('Texture#NPOT Workaround: texture creation', t => {
-  // Create NPOT texture with no parameters
-  let texture = webglDevice.createTexture({data: null, width: 500, height: 512});
-  t.ok(texture instanceof Texture, 'Texture construction successful');
-
-  // Default parameters should be changed to supported NPOT parameters.
-  let minFilter = getParameter(texture, 'minFilter');
-  t.equals(minFilter, GL.LINEAR, 'NPOT texture min filter is set to LINEAR');
-  let wrapS = getParameter(texture, 'addressModeU');
-  t.equals(wrapS, 'clamp-to-edge', 'NPOT texture wrap_s is set to CLAMP_TO_EDGE');
-  let wrapT = getParameter(texture, 'addressModeV');
-  t.equals(wrapT, 'clamp-to-edge', 'NPOT texture wrap_t is set to CLAMP_TO_EDGE');
-
-  const parameters = {
-    ['minFilter']: 'nearest',
-    ['addressModeU']: 'repeat',
-    ['addressModeV']: 'mirrored-repeat'
-  };
-
-  // Create NPOT texture with parameters
-  texture = webglDevice.createTexture({
-    data: null,
-    width: 512,
-    height: 600,
-    parameters
-  });
-  t.ok(texture instanceof Texture, 'Texture construction successful');
-
-  // Above parameters should be changed to supported NPOT parameters.
-  minFilter = getParameter(texture, 'minFilter');
-  t.equals(minFilter, 'nearest', 'NPOT texture min filter is set to NEAREST');
-  wrapS = getParameter(texture, 'addressModeU');
-  t.equals(wrapS, 'clamp-to-edge', 'NPOT texture wrap_s is set to CLAMP_TO_EDGE');
-  wrapT = getParameter(texture, 'addressModeV');
-  t.equals(wrapT, 'clamp-to-edge', 'NPOT texture wrap_t is set to CLAMP_TO_EDGE');
+      if (device.info.type === 'webgl') {
+        t.deepEquals(
+          device.readPixelsToArrayWebGL(texture),
+          new Uint8Array([0, 0, 0, 255, 0, 0, 0, 255]),
+          `${device.info.type} Pixels are then set correctly`
+        );
+      } else {
+        t.pass('WebGPU copyExternalImage test cannot yet read pixels');
+      }
+    }
+  }
 
   t.end();
 });
 
-test.skip('Texture#NPOT Workaround: setParameters', t => {
-  // Create NPOT texture
-  const texture = webglDevice.createTexture({data: null, width: 100, height: 100});
-  t.ok(texture instanceof Texture, 'Texture construction successful');
-
-  const invalidNPOTParameters = {
-    ['minFilter']: 'linear',
-    ['mipmapFilter']: 'nearest',
-    ['addressModeU']: 'mirrored-repeat',
-    ['addressModeV']: 'repeat'
-  };
-  texture.setParameters(invalidNPOTParameters);
-
-  // Above parameters should be changed to supported NPOT parameters.
-  const minFilter = getParameter(texture, 'minFilter');
-  t.equals(minFilter, 'linear', 'NPOT texture min filter is set to LINEAR');
-  const wrapS = getParameter(texture, 'addressModeU');
-  t.equals(wrapS, 'clamp-to-edge', 'NPOT texture wrap_s is set to CLAMP_TO_EDGE');
-  const wrapT = getParameter(texture, 'addressModeV');
-  t.equals(wrapT, 'clamp-to-edge', 'NPOT texture wrap_t is set to CLAMP_TO_EDGE');
-
-  t.end();
-});
-
-test.skip('WebGL2#Texture setImageData', t => {
+test.skip('Texture#setImageData', t => {
   let data;
 
   // data: null
   const texture = webglDevice.createTexture({
-    data: null,
     width: 2,
     height: 1,
     format: GL.RGBA32F,
@@ -318,30 +259,8 @@ test.skip('WebGL2#Texture setImageData', t => {
 
   // data: typed array
   data = new Float32Array([0.1, 0.2, -3, -2, 0, 0.5, 128, 255]);
-  texture.setImageData({data});
+  texture.copyExternalImage({data});
   t.deepEquals(readPixelsToArray(texture), data, 'Pixels are set correctly');
-
-  // data: buffer
-  data = new Float32Array([21, 0.82, 0, 1, 0, 255, 128, 3.333]);
-  const buffer = webglDevice.createByffer({data, accessor: {size: 4, type: GL.FLOAT}});
-  texture.setImageData({data: buffer});
-  t.deepEquals(readPixelsToArray(texture), data, 'Pixels are set correctly');
-
-  // data: canvas
-  if (typeof document !== 'undefined') {
-    const canvas = document.createElement('canvas');
-    canvas.width = 2;
-    canvas.height = 1;
-    const ctx = canvas.getContext('2d');
-    ctx.fillRect(0, 0, 2, 1);
-    texture.setImageData({data: canvas});
-    t.deepEquals(
-      readPixelsToArray(texture),
-      new Float32Array([0, 0, 0, 1, 0, 0, 0, 1]),
-      'Pixels are set correctly'
-    );
-  }
-
   t.end();
 });
 
@@ -420,49 +339,21 @@ test.skip('WebGL2#Texture generateMipmap', t => {
   t.end();
 });
 
-// Shared with texture*.spec.js
-export function testSamplerParameters({t, texture, parameters}) {
-  for (const parameterName in parameters) {
-    const values = parameters[parameterName];
-    const parameter = Number(parameterName);
-    for (const valueName in values) {
-      const value = Number(valueName);
-      texture.setParameters({
-        [parameter]: value
-      });
-      const name = texture.constructor.name;
-      const newValue = getParameter(texture, parameter);
-      t.equals(
-        newValue,
-        value,
-        `${name}.setParameters({[${device.getGLKey(parameter)}]: ${device.getGLKey(
-          value
-        )}}) read back OK`
+// NON-2D TEXTURES
+
+test('Texture(dimension)#construct/delete', async t => {
+  for (const device of await getTestDevices()) {
+    for (const dimension of ['3d', '2d-array', 'cube']) {
+      const texture = webglDevice.createTexture({dimension});
+      t.equal(texture.dimension, dimension, `${device.info.type} Texture construction successful`);
+      t.equal(
+        texture.view.props.dimension,
+        dimension,
+        `${device.info.type} Texture view construction successful`
       );
+      texture.destroy();
     }
   }
-}
-
-// 2D TEXTURES
-
-test.skip('Texture#construct/delete', t => {
-  t.throws(
-    // @ts-expect-error
-    () => new Texture(),
-    /.*WebGLRenderingContext.*/,
-    'Texture throws on missing gl context'
-  );
-
-  const texture = webglDevice.createTexture();
-  t.ok(texture instanceof Texture, 'Texture construction successful');
-
-  t.comment(JSON.stringify(texture.getParameters({keys: true})));
-
-  texture.destroy();
-  t.ok(texture instanceof Texture, 'Texture delete successful');
-
-  texture.destroy();
-  t.ok(texture instanceof Texture, 'Texture repeated delete successful');
 
   t.end();
 });
@@ -582,12 +473,26 @@ test.skip('WebGL#Texture3D construct/delete', t => {
   t.end();
 });
 
-// HELPERS
+test.skip('Texture#setParameters', t => {
+  const texture = webglDevice.createTexture({});
+  t.ok(texture instanceof Texture, 'Texture construction successful');
 
-function getParameter(texture: Texture, pname: number): any {
-  const webglTexture = texture as WEBGLTexture;
-  webglTexture.gl.bindTexture(webglTexture.target, webglTexture.handle);
-  const value = webglTexture.gl.getTexParameter(webglTexture.target, pname);
-  webglTexture.gl.bindTexture(webglTexture.target, null);
-  return value;
-}
+  testSamplerParameters({t, texture, sampler: SAMPLER_PARAMETERS});
+
+  /*
+  // Bad tests
+  const parameter = GL.TEXTURE_MAG_FILTER;
+  const value = GL.LINEAR_MIPMAP_LINEAR;
+  texture.setParameters({
+    [parameter]: value
+  });
+  const newValue = getParameter(texture, GL.TEXTURE_MAG_FILTER);
+  t.equals(newValue, value,
+    `Texture.setParameters({[${device.getGLKey(parameter)}]: ${device.getGLKey(value)}})`);
+  */
+
+  texture.destroy();
+  t.ok(texture instanceof Texture, 'Texture delete successful');
+
+  t.end();
+});