+++ modules/test-utils/src/null-device/resources/null-texture.ts
@@ -2,7 +2,13 @@
 // SPDX-License-Identifier: MIT
 // Copyright (c) vis.gl contributors
 
-import type {TextureProps, Sampler, SamplerProps, TextureViewProps} from '@luma.gl/core';
+import type {
+  TextureProps,
+  Sampler,
+  SamplerProps,
+  TextureViewProps,
+  CopyExternalImageOptions
+} from '@luma.gl/core';
 import type {
   Texture1DData,
   Texture2DData,
@@ -74,21 +80,8 @@ export class NullTexture extends Texture {
   }
 
   initialize(props: TextureProps = {}): this {
-    const data = props.data;
-
-    if (data instanceof Promise) {
-      data.then(resolvedImageData =>
-        this.initialize(
-          Object.assign({}, props, {
-            pixels: resolvedImageData,
-            data: resolvedImageData
-          })
-        )
-      );
-      return this;
-    }
-
-    this.setImageData(props);
+    // const data = props.data;
+    // this.setImageData(props);
 
     this.setSampler(props.sampler);
 
@@ -112,17 +105,17 @@ export class NullTexture extends Texture {
     return this;
   }
 
-  setImageData(options: {data?: any; width?: number; height?: number}) {
+  copyExternalImage(options: CopyExternalImageOptions): {width: number; height: number} {
     this.trackDeallocatedMemory('Texture');
 
-    const {data} = options;
+    const {image: data} = options;
 
-    if (data && data.byteLength) {
-      this.trackAllocatedMemory(data.byteLength, 'Texture');
-    } else {
-      const bytesPerPixel = 4;
-      this.trackAllocatedMemory(this.width * this.height * bytesPerPixel, 'Texture');
-    }
+    // if (data && data.byteLength) {
+    //   this.trackAllocatedMemory(data.byteLength, 'Texture');
+    // } else {
+    const bytesPerPixel = 4;
+    this.trackAllocatedMemory(this.width * this.height * bytesPerPixel, 'Texture');
+    // }
 
     const width = options.width ?? (data as ImageBitmap).width;
     const height = options.height ?? (data as ImageBitmap).height;
@@ -130,13 +123,6 @@ export class NullTexture extends Texture {
     this.width = width;
     this.height = height;
 
-    return this;
-  }
-
-  setSubImageData(options: {data: any; width?: number; height?: number; x?: number; y?: number}) {
-    // const {data, x = 0, y = 0} = options;
-    // const width = options.width ?? (data as ImageBitmap).width;
-    // const height = options.height ?? (data as ImageBitmap).height;
-    // assert(width + x <= this.width && height + y <= this.height);
+    return {width, height};
   }
 }